meta-intel: lms7: handle different host setups
From similar patch for lms8 (fc9cba84):

The configure script looks at the *host* environment to decide where
to install the init script, so it's location at packaging time can
change.

(From meta-intel rev: e7f77e4dc4490a10a597d3da7e85d6650ccb6fc3)

+++ meta-intel/common/recipes-bsp/amt/lms7_7.1.20.bb
@@ -34,7 +34,13 @@ addtask unpack2 after do_unpack before do_patch
 do_install_append () {
 	mv ${D}/${sbindir}/lms ${D}/${sbindir}/lms7
 	install -d ${D}${sysconfdir}/init.d
-	mv ${D}${sysconfdir}/rc.d/init.d/lms ${D}${sysconfdir}/init.d/${INITSCRIPT_NAME}
+	# The configure script looks at the host to decide where to put init
+	# scripts, so move it at the same time as renaming it.
+	if test -f ${D}${sysconfdir}/rc.d/init.d/lms ; then
+		mv ${D}${sysconfdir}/rc.d/init.d/lms ${D}${sysconfdir}/init.d/${INITSCRIPT_NAME}
+	else
+		mv ${D}${sysconfdir}/init.d/lms ${D}${sysconfdir}/init.d/${INITSCRIPT_NAME}
+	fi
 	sed -i 's/^NAME=lms/NAME=lms7/' ${D}${sysconfdir}/init.d/${INITSCRIPT_NAME}
 	rmdir ${D}${datadir} || :
 }
